Fixeado combo boxes con "Cualquiera"

+++ src/main/java/DAO/DAOBdInmueble.java
@@ -52,7 +52,7 @@ public class DAOBdInmueble implements InmuebleDAO{
     }
 
     @Override
-    public List<Inmueble> listAllByPropietario(long idPropietario) {
+    public List<Inmueble> listAllByPropietario(int idPropietario) {
 
         EntityManager manager = Conexion.emf.createEntityManager();
         @SuppressWarnings("unchecked")
@@ -62,7 +62,7 @@ public class DAOBdInmueble implements InmuebleDAO{
     }
 
     @Override
-    public List<Inmueble> listAllByPropietario(long idPropietario, int inicio, int fin) {
+    public List<Inmueble> listAllByPropietario(int idPropietario, int inicio, int fin) {
         EntityManager manager = Conexion.emf.createEntityManager();
         @SuppressWarnings("unchecked")
         List<Inmueble> lista = (List<Inmueble>) manager.createQuery("From Inmueble as i Where i.estado!='BAJA' AND i.propietarioInmueble.id= :idprop").setParameter("idprop",idPropietario).setMaxResults(fin - inicio + 1).setFirstResult(inicio - 1).getResultList();
